+++ src/app/components/advertisement-index/advertisement-index.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+import { AdvertisementService } from '../../services/advertisement.service';
+
+@Component({
+  selector: 'app-advertisement-index',
+  templateUrl: './advertisement-index.component.html',
+  styleUrls: ['./advertisement-index.component.scss']
+})
+export class AdvertisementIndexComponent implements OnInit {
+
+  advertisements$;
+
+  constructor(private advertisementService: AdvertisementService) { }
+
+  ngOnInit() {
+    this.advertisements$ = this.advertisementService.getAllAdvertisements();
+  }
+
+}